+++ src/main/java/jpos/loader/simple/SimpleServiceManager.java
@@ -116,7 +116,7 @@ public class SimpleServiceManager implements JposServiceManager
     private ProfileFactory getProfileFactory()
     {
 		if( profileFactory == null )
-			profileFactory = new XercesProfileFactory();
+			profileFactory = new DefaultProfileFactory();
 		
 		return profileFactory;
     }